fix sonar issue: auth doubletten in component
