+++ frontend/src/model/Credentials.ts
@@ -1,0 +1,4 @@
+export type Credentials = {
+    username:string;
+    password:string;
+}
